+++ src/main/java/baseball/domain/vo/Point.java
@@ -1,4 +1,4 @@
-package baseball.vo;
+package baseball.domain.vo;
 
 public class Point {
     private int strike;
